Story:  Add intelligent auto-dummy support for testing. [2076]
git-svn-id: https://geekscape.org/svn/boost/trunk@2362 05db6a09-9903-0410-bedf-ba9bc4063e48

+++ src/test/java/au/net/netstorm/boost/time/core/DefaultClockAtomicTest.java
@@ -28,7 +28,7 @@ public final class DefaultClockAtomicTest extends InteractionTestCase implements
 
     public void testGetTime() {
         expect.oneCall(timeLord, expected, "get", millis);
-        TimePoint actual = subject.getTime(millis);
+        TimePoint actual = subject.getTime(millis.longValue());
         assertEquals(expected, actual);
     }
